requirements.txt
Flask>=2.2
Flask-SQLAlchemy==3.0.5
PyMySQL>=1.0
python-dotenv>=0.21